Create new TextBuffer class
